Implemented `block` method of target
The block function replaces the Self::func calls in if, while, and for.
This drastically reduces allocations, and HUGELY speeds up runtime. Now,
the virtual machine isnt duplicated 3 times every if statement. The rust
version of block is weird now because Value::function implicitly duplicates
the machine it gets. Next, xmachine needs to remove that feature so xassembler
can choose when to duplicate entirely. The rust version still gets a massive
speed boost as well even tho the impl is weird, tho